requirements.txt
fastapi~=0.112.2
sqlalchemy~=2.0.33
uvicorn
pydantic~=2.8.2
starlette~=0.38.4
bcrypt~=4.2.0
passlib~=1.7.4
python-jose~=3.3.0
python-dotenv~=1.0.1
python-multipart~=0.0.9